Reveal canvas help on hover like the notifications bell
The bottom-left canvas help previously required a click to toggle its
popover. Extract it into a dedicated CanvasHelpButton that mirrors the
toolbar notifications bell: hovering reveals the popover, moving away
dismisses it after a short grace period, and clicking pins it open.
